touchups

+++ dev-util/ddd/ddd-3.3.1-r3.ebuild
@@ -1,13 +1,10 @@
 # Copyright 1999-2003 Gentoo Technologies, Inc.
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/dev-util/ddd/ddd-3.3.1-r3.ebuild,v 1.6 2003/03/30 00:36:40 liquidx Exp $
-
-IUSE=""
+# $Header: /var/cvsroot/gentoo-x86/dev-util/ddd/ddd-3.3.1-r3.ebuild,v 1.7 2003/04/23 16:13:27 vapier Exp $
 
 inherit eutils
 
-S="${WORKDIR}/${P}"
-DESCRIPTION="GNU DDD is a graphical front-end for command-line debuggers"
+DESCRIPTION="graphical front-end for command-line debuggers"
 HOMEPAGE="http://www.gnu.org/software/ddd"
 SRC_URI="ftp://ftp.easynet.be/gnu/ddd/${P}.tar.gz
 	ftp://ftp.easynet.be/gnu/ddd/${P}-html-manual.tar.gz"
@@ -19,7 +16,6 @@ KEYWORDS="x86 ppc sparc alpha"
 DEPEND="virtual/x11
 	>=sys-devel/gdb-4.16
 	virtual/motif"
-
 
 src_unpack() {
 	unpack ${A}
@@ -41,7 +37,7 @@ src_compile() {
 	emake || die
 }
 
-src_install () {
+src_install() {
 	dodir /usr/lib
 	# If using internal libiberty.a, need to pass
 	# $tooldir to 'make install', else we get
@@ -50,17 +46,13 @@ src_install () {
 	einstall tooldir=${D}/usr || die
 
 	# This one is from binutils
-	if [ -f ${D}/usr/lib/libiberty.a ]
-	then
-		rm -f ${D}/usr/lib/libiberty.a
-	fi
+	[ -f ${D}/usr/lib/libiberty.a ] && rm -f ${D}/usr/lib/libiberty.a
 	# Remove empty dir ...
 	rmdir ${D}/usr/lib || :
-	
+
 	mv ${S}/doc/README ${S}/doc/README-DOC
 	dodoc ANNOUNCE AUTHORS BUGS COPYING* CREDITS INSTALL NEWS* NICKNAMES \
 		OPENBUGS PROBLEMS README* TIPS TODO
-	
+
 	mv ${S}/doc/* ${D}/usr/share/doc/${PF}
 }
-
